backend/pyproject.toml
[tool.poetry]
name = "account-research-bot"
version = "0.1.0"
description = ""
authors = ["Carlos Castro <carlos.castro@mongodb.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
python-dotenv = "^1.0.1"
mongo-parent-document-retrieval = [
    {path = "packages/mongo-parent-document-retrieval", develop = true, platform = 'linux'},
    {path = "packages\\mongo-parent-document-retrieval", develop = true, platform = 'win32'}
]

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
